implement event-resgister

+++ src/main/java/com/jesus/challenge/bvc/models/dto/EventRegisterDto.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,7 +15,7 @@ public class EventRegisterDto {
     private Long id;
     private Long date;
     private Integer amount;
-    private Integer cost;
+    private BigDecimal cost;
     private Platform platform;
     private Event event;
 }
